some updates after data massive input.

+++ indicators/adr_atr.py
@@ -23,7 +23,7 @@ def atr(instrument_id: int, period: int = 30) -> pd.DataFrame:
     if len(df) < period:
         return pd.DataFrame()
     col = f"atr_{period}d"
-    df[col] = talib.ATR(df["high"].values, df["low"].values, df["close"].values, timeperiod=period)
+    df[col] = talib.ATR(df["high"].astype(float).values, df["low"].astype(float).values, df["close"].astype(float).values, timeperiod=period)
     result = df[["date", col]].dropna(subset=[col]).copy()
     result["instrument_id"] = instrument_id
     result["indicator_name"] = "atr"
