+++ indicators/obv.py
@@ -7,7 +7,7 @@ def obv(instrument_id: int, period: int = 100) -> pd.DataFrame:
     df = load_quotes(instrument_id)
     if df.empty:
         return pd.DataFrame()
-    obv_values = talib.OBV(df["close"].values, df["volume"].values)
+    obv_values = talib.OBV(df["close"].astype(float).values, df["volume"].astype(float).values)
     col = f"obv_{period}d"
     df[col] = talib.SMA(obv_values, timeperiod=period)
     result = df[["date", col]].dropna(subset=[col]).copy()
